backend/requirements.txt
transformers>=4.50.0
accelerate>=0.26.0
bitsandbytes>=0.41.0
torch>=2.0.0
Pillow>=9.0.0
fastapi
uvicorn
python-multipart
